chore: separating Api class
