+++ client/src/utility-classes/Api.js
@@ -1,38 +1,6 @@
-class Api {
-  static instance
-  static jsonContentHeader = { 'Content-Type': 'application/json' }
+import HTTPMethodsHandler from './HTTPMethodsHandler'
 
-  constructor(apiBase) {
-    if (Api.instance) {
-      return Api.instance
-    }
-    this.base = apiBase
-    Api.instance = this
-  }
-
-  static init(apiBase) {
-    new Api(apiBase)
-  }
-
-  static async fetch(urlPart, method, data) {
-    const url = `/${this.instance.base}/${urlPart}`
-
-    const params = { headers: this.jsonContentHeader, method }
-    if (method === 'POST' || method === 'PUT')
-      params.body = JSON.stringify(data)
-
-    return await fetch(url, params).then(async (res) => {
-      if (!res.ok) {
-        throw new Error(`Data retrieval error. Status: ${res.status}`)
-      }
-      return await res.json()
-    })
-  }
-
-  static async get(urlPart) {
-    return await Api.fetch(urlPart, 'GET')
-  }
-
+class Api extends HTTPMethodsHandler {
   static async getLength(routeForContent, param) {
     return await Api.get(`${routeForContent}/length/${param || ''}`)
   }
@@ -76,10 +44,6 @@ class Api {
     return await Api.get('groups/list-of-titles')
   }
 
-  static async delete(urlPart) {
-    return await Api.fetch(urlPart, 'DELETE')
-  }
-
   static async deleteEntity(entity, param) {
     return await Api.fetch(`${entity}/${param}`, 'DELETE')
   }
@@ -90,14 +54,6 @@ class Api {
 
   static async deleteGroup(param) {
     return await Api.deleteEntity('groups', param)
-  }
-
-  static async post(urlPart, data) {
-    return await Api.fetch(urlPart, 'POST', data)
-  }
-
-  static async put(urlPart, data) {
-    return await Api.fetch(urlPart, 'PUT', data)
   }
 
   static async addNewEntity(entity, data) {
